io7m-jlog-core/src/main/java/com/io7m/jlog/Log.java
/*
 * Copyright © 2014 <code@io7m.com> http://io7m.com
 * 
 * Permission to use, copy, modify, and/or distribute this software for any
 * purpose with or without fee is hereby granted, provided that the above
 * copyright notice and this permission notice appear in all copies.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
 * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
 * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
 * SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
 * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
 * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
 * IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 */

package com.io7m.jlog;

import com.io7m.jequality.annotations.EqualityReference;
import com.io7m.junreachable.UnreachableCodeException;

/**
 * <p>
 * Constructor functions for obtaining {@link LogType} implementations.
 * </p>
 */

@EqualityReference public final class Log
{
  /**
   * Construct a new root log interface, logging to the given destination.
   * 
   * @param policy
   *          The log policy
   * @param destination
   *          The log destination
   * @return A new log interface
   */

  public static LogType newLog(
    final LogPolicyType policy,
    final String destination)
  {
    return new LogRoot(policy, destination);
  }

  private Log()
  {
    throw new UnreachableCodeException();
  }
}
